Remove fetch and add support for PHP7.4
