+++ composer.json
@@ -5,7 +5,6 @@
     "keywords": ["supabase","postgrest", "postgrest-php"],
     "homepage": "https://github.com/supabase-community/postgrest-php",
     "autoload": {
-	 "psr-0": { "": "src/" },
         "psr-4": {
             "Supabase\\Postgrest\\": "src/"
         }
@@ -29,13 +28,14 @@
             "email": "adolfo@zerocopylabs.com"
         }
     ],
-    "require-dev": {
-        "phpunit/phpunit": "9.6",
-        "allure-framework/allure-phpunit": "1.5"
-    },
     "require": {
+        "php": "^7.4",
         "guzzlehttp/guzzle": "7.5",
         "vlucas/phpdotenv": "5.5",
         "spatie/url": "2.2"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "9.6",
+        "allure-framework/allure-phpunit": "1.5"
     }
 }
